+++ graph-commons/src/main/scala/ch/datascience/http/rest/paging/model.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2019 Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ch.datascience.http.rest.paging
+
+import ch.datascience.tinytypes.constraints.PositiveInt
+import ch.datascience.tinytypes.{IntTinyType, TinyTypeFactory}
+
+object model {
+
+  final class Page private (val value: Int) extends AnyVal with IntTinyType
+  implicit object Page extends TinyTypeFactory[Page](new Page(_)) with PositiveInt {
+    val first: Page = Page(1)
+  }
+
+  final class PerPage private (val value: Int) extends AnyVal with IntTinyType
+  implicit object PerPage extends TinyTypeFactory[PerPage](new PerPage(_)) with PositiveInt {
+    val default: PerPage = PerPage(20)
+  }
+}
