+++ graph-commons/src/test/scala/ch/datascience/http/rest/paging/PagingRequestSpec.scala
@@ -16,37 +16,35 @@
  * limitations under the License.
  */
 
-package ch.datascience.http.rest
+package ch.datascience.http.rest.paging
 
 import cats.data.{NonEmptyList, Validated}
 import cats.implicits._
 import ch.datascience.generators.CommonGraphGenerators._
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators._
-import ch.datascience.http.rest.Paging.PagingRequest
-import ch.datascience.http.rest.Paging.PagingRequest.{Page, PerPage}
-import ch.datascience.tinytypes.constraints.PositiveInt
+import ch.datascience.http.rest.paging.model.{Page, PerPage}
 import org.http4s.ParseFailure
 import org.scalacheck.Gen
 import org.scalatest.Matchers._
 import org.scalatest.WordSpec
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
-class PagingSpec extends WordSpec with ScalaCheckPropertyChecks {
+class PagingRequestSpec extends WordSpec with ScalaCheckPropertyChecks {
 
   "page" should {
 
     "decode a valid page query parameter" in {
-      forAll { page: PagingRequest.Page =>
+      forAll { page: Page =>
         Map("page" -> Seq(page.toString)) match {
-          case Paging.PagingRequest.Decoders.page(actual) => actual shouldBe Some(Validated.validNel(page))
+          case PagingRequest.Decoders.page(actual) => actual shouldBe Some(Validated.validNel(page))
         }
       }
     }
 
     "fail to decode a non-int page query parameter" in {
       Map("page" -> Seq("abc")) match {
-        case Paging.PagingRequest.Decoders.page(actual) =>
+        case PagingRequest.Decoders.page(actual) =>
           actual shouldBe Some(Validated.invalidNel {
             ParseFailure("'abc' not a valid Page number", "")
           })
@@ -55,7 +53,7 @@ class PagingSpec extends WordSpec with ScalaCheckPropertyChecks {
 
     "fail to decode a non-positive page query parameter" in {
       Map("page" -> Seq("0")) match {
-        case Paging.PagingRequest.Decoders.page(actual) =>
+        case PagingRequest.Decoders.page(actual) =>
           actual shouldBe Some(Validated.invalidNel {
             ParseFailure("'0' not a valid Page number", "")
           })
@@ -64,7 +62,7 @@ class PagingSpec extends WordSpec with ScalaCheckPropertyChecks {
 
     "return None when no page query parameter" in {
       Map.empty[String, List[String]] match {
-        case Paging.PagingRequest.Decoders.page(actual) => actual shouldBe None
+        case PagingRequest.Decoders.page(actual) => actual shouldBe None
       }
     }
   }
@@ -72,16 +70,16 @@ class PagingSpec extends WordSpec with ScalaCheckPropertyChecks {
   "perPage" should {
 
     "decode a valid per_page query parameter" in {
-      forAll { perPage: PagingRequest.PerPage =>
+      forAll { perPage: PerPage =>
         Map("per_page" -> Seq(perPage.toString)) match {
-          case Paging.PagingRequest.Decoders.perPage(actual) => actual shouldBe Some(Validated.validNel(perPage))
+          case PagingRequest.Decoders.perPage(actual) => actual shouldBe Some(Validated.validNel(perPage))
         }
       }
     }
 
     "fail to decode a non-int per_page query parameter" in {
       Map("per_page" -> Seq("abc")) match {
-        case Paging.PagingRequest.Decoders.perPage(actual) =>
+        case PagingRequest.Decoders.perPage(actual) =>
           actual shouldBe Some(Validated.invalidNel {
             ParseFailure("'abc' not a valid PerPage number", "")
           })
@@ -90,7 +88,7 @@ class PagingSpec extends WordSpec with ScalaCheckPropertyChecks {
 
     "fail to decode a non-positive per_page query parameter" in {
       Map("per_page" -> Seq("0")) match {
-        case Paging.PagingRequest.Decoders.perPage(actual) =>
+        case PagingRequest.Decoders.perPage(actual) =>
           actual shouldBe Some(Validated.invalidNel {
             ParseFailure("'0' not a valid PerPage number", "")
           })
@@ -99,20 +97,8 @@ class PagingSpec extends WordSpec with ScalaCheckPropertyChecks {
 
     "return None when no per_page query parameter" in {
       Map.empty[String, List[String]] match {
-        case Paging.PagingRequest.Decoders.perPage(actual) => actual shouldBe None
+        case PagingRequest.Decoders.perPage(actual) => actual shouldBe None
       }
-    }
-  }
-
-  "Page" should {
-    "be a PositiveInt constrained" in {
-      Page shouldBe a[PositiveInt]
-    }
-  }
-
-  "PerPage" should {
-    "be a PositiveInt constrained" in {
-      PerPage shouldBe a[PositiveInt]
     }
   }
