+++ knowledge-graph/src/main/scala/ch/datascience/knowledgegraph/datasets/rest/DatasetsFinder.scala
@@ -20,7 +20,7 @@ package ch.datascience.knowledgegraph.datasets.rest
 
 import cats.effect.{ContextShift, IO, Timer}
 import ch.datascience.graph.model.datasets.{Description, Identifier, Name, PublishedDate}
-import ch.datascience.http.rest.Paging.PagingRequest
+import ch.datascience.http.rest.paging.PagingRequest
 import ch.datascience.knowledgegraph.datasets.CreatorsFinder
 import ch.datascience.knowledgegraph.datasets.model.DatasetPublishing
 import ch.datascience.knowledgegraph.datasets.rest.DatasetsFinder.DatasetSearchResult
