+++ knowledge-graph/src/main/scala/ch/datascience/knowledgegraph/datasets/rest/DatasetsSearchEndpoint.scala
@@ -26,7 +26,7 @@ import ch.datascience.controllers.ErrorMessage
 import ch.datascience.controllers.InfoMessage._
 import ch.datascience.graph.config.RenkuBaseUrl
 import ch.datascience.http.rest.Links.{Href, Link, Rel, _links}
-import ch.datascience.http.rest.Paging.PagingRequest
+import ch.datascience.http.rest.paging.PagingRequest
 import ch.datascience.knowledgegraph.datasets.CreatorsFinder
 import ch.datascience.knowledgegraph.datasets.model.{DatasetCreator, DatasetPublishing}
 import ch.datascience.logging.{ApplicationLogger, ExecutionTimeRecorder}
